feat: add authentication to HTTP requests

+++ fastapi_tutorial/routers/todos.py
@@ -7,6 +7,7 @@ from fastapi_tutorial.models import Todos
 from fastapi_tutorial.database import SessionLocal
 from starlette import status
 from pydantic import BaseModel, Field
+from .auth import get_current_user
 
 # uvicorn fastapi_tutorial.main:app --reload
 router = APIRouter()
@@ -34,6 +35,7 @@ def get_db():
         db.close()
 
 db_dependency = Annotated[Session, Depends(get_db)]
+user_dependency = Annotated[dict, Depends(get_current_user)]
 
 class TodoRequest(BaseModel):
     # don't have to specify id here because it is automatically added in db
@@ -43,30 +45,68 @@ class TodoRequest(BaseModel):
     complete: bool
 
 @router.get("/", status_code=status.HTTP_200_OK)
-async def read_all(db: db_dependency):
-    return db.query(Todos).all()
+async def read_all(
+    user: user_dependency,
+    db: db_dependency
+):
+    if user is None:
+        raise HTTPException(status_code=401, detail="Authentication failed.")
+    
+    return db.query(Todos).filter(Todos.owner_id == user.get("id")).all()
 
 @router.get("/todo/{todo_id}", status_code=status.HTTP_200_OK)
-async def read_todo(db: db_dependency, todo_id: int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def read_todo(
+    user: user_dependency,
+    db: db_dependency, 
+    todo_id: int = Path(gt=0)
+):
+    if user is None:
+        raise HTTPException(status_code=401, detail="Authentication failed.")
+    
+    todo_model = (
+        db
+        .query(Todos)
+        .filter(Todos.id == todo_id)
+        .filter(Todos.owner_id == user.get("id"))
+        .first()
+    )
     if todo_model:
         return todo_model
     raise HTTPException(status_code=404, detail="Todo not found.")
 
 @router.post("/todo", status_code=status.HTTP_201_CREATED)
-async def create_table(db: db_dependency, todo_request: TodoRequest):
-    todo_model = Todos(**todo_request.model_dump())
+async def create_todo(
+    user: user_dependency,
+    db: db_dependency, 
+    todo_request: TodoRequest
+):
+    if user is None:
+        raise HTTPException(status_code=401, detail="Authentication failed.")
+    todo_model = Todos(
+        **todo_request.model_dump(),
+        owner_id=user.get("id")
+    )
 
     db.add(todo_model)
     db.commit()
 
 @router.put("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
 async def update_todo(
+    user: user_dependency,
     db: db_dependency,
     todo_request: TodoRequest,
     todo_id: int = Path(gt=0),
 ):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+    if user is None:
+        raise HTTPException(status_code=401, detail="Authentication failed.")
+    
+    todo_model = (
+        db
+        .query(Todos)
+        .filter(Todos.id == todo_id)
+        .filter(Todos.owner_id == user.get("id"))
+        .first()
+    )
     if todo_model is None:
         raise HTTPException(status_code=404, detail="Todo not found.")
 
@@ -79,9 +119,22 @@ async def update_todo(
     db.commit()
 
 @router.delete("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_todo(db: db_dependency, todo_id: int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def delete_todo(
+    user: user_dependency,
+    db: db_dependency, 
+    todo_id: int = Path(gt=0)
+):
+    if user is None:
+        raise HTTPException(status_code=401, detail="Authentication failed.")
+    
+    todo_model = (
+        db
+        .query(Todos)
+        .filter(Todos.id == todo_id)
+        .filter(Todos.owner_id == user.get("id"))
+        .first()
+    )
     if todo_model is None:
         raise HTTPException(status_code=404, detail="Todo not found")
-    db.query(Todos).filter(Todos.id == todo_id).delete()
+    db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get("id")).delete()
     db.commit()
